feat: place table datas into manifest, update them together (#863)
## Which issue does this PR close?

Closes #
Part of #799 

## Rationale for this change
Now, we update `TableData` and store its wal seperately. The order of
two operations above is maintained by developer, that will lead to a big
bug if developer is not so careful when modifying related codes.
<!---
Why are you proposing this change? If this is already explained clearly
in the issue, then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

## What changes are included in this PR?
+ Place table datas into manifest.
+ Update it both in memory and storage in the same call.
<!---
There is no need to duplicate the description in the issue here, but it
is sometimes worth providing a summary of the individual changes in this
PR to help reviewers understand the structure.
-->

## Are there any user-facing changes?
None.
<!---
Please mention if:

- there are user-facing changes that need to update the documentation or
configuration.
- this is a breaking change to public APIs
-->

## How does this change test
Test by ut.
<!-- 
Please describe how you test this change (like by unit test case,
integration test or some other ways) if this change has touched the
code.
-->