fix: write cancel when flush pending write queue (#940)
## Related Issues
Closes #

## Detailed Changes
Currently, the writer for the first request in the pending write queue
is responsible for flushing all the writes in the queue, but it can't be
ensured to always trigger flush because it may be cancelled when await
for the serial exec lock. And this change set replaces the
`try_join_all` in the rpc services to avoid such cancellation.

## Test Plan 
Test it manually.