feat: avoid prefetching all sst streams at once (#1069)
## Rationale
Close #959

The current procedure to fetch sst data is started immediately, and it
may lead to high memory consumption when massive concurrent queries
reaches.

## Detailed Changes
Introduce a prefetchable stream to replace the normal stream, with which
a way to trigger the fetching data is provided and when to trigger it
can also be determined by the caller.

## Test Plan
Test locally with massive concurrent queries. The memory increment is much slower after this patch.