chore: remove the codes about the reverse reading (#1116)
## Rationale
Currently, the read in a reverse order of primary key is not supported
yet. Remove the codes about that.

## Detailed Changes
Remove the codes about the reverse read during plan stage and reading
sst stage, while the ability to scan memtable in reverse order is kept
for possible future use.

## Test Plan
Existing tests.