feat: improved partitioned table tests powered by tsbs (#1195)
## Rationale
Close #1191 

The integration test for partitioned table is too simple, now. We just
insert data of poor amount and query them...
In this, tsbs is introudced for big data amount test for partitioned
table.

## Detailed Changes
+ Introduce tsbs and a pre-generated dataset for testing.
+ We don't want to remove and rebuild components(e.g. ceresmeta) on
sometimes, and I added an option to control it in this pr, usage:
```
UPDATE_REPOS_TO_LATEST=false make run-dist-query
```

## Test Plan
It is the test itself.