feat: support random partition rule (#1193)
## Rationale
Currently, hash/key rule are the only supported rules for partition
table. However, it will lead to small write batches if the user tries to
use primary key or all tags as the partition key for a uniform
distribution. And it is known that the small write batches hurt write
performance badly.

Actually, for a partition table in `append` mode with primary key or all
tags as partition key, most of the queries to this table will fan out to
all the partitions because the partition key won't be provided in most
cases. So what is the gain from the cost (the massive small write
batches) to do such complex partition?

It will work much better if a simple partition rule is supported, by
which a random partition is chosen for a given write request.

## Detailed Changes
- Introduce `RandomPartitionRule`
- Add a new method called `validate_create_table` to `TableEngine` trait

## Test Plan
Add unit test to apply `RandomPartitionRule` for a partition table in
different mode.