feat: add deploy mode (#247)
## Rationale
In the current implementation, every time the CeresDB node changes, it
will cause large-scale scheduling in the cluster. However, in the actual
situation, we hope to control the affected range of the cluster when
deploy and debug. Therefore, we need to add a deploy mode switch to
control cluster scheduling.

## Detailed Changes
* Add `DeployMode` switch in `RebalancedShardScheduler`, shard topology
will be locked when deploy mode is true.
* Add some API to control `DeployMode`.

## Test Plan
Pass all unit tests and test `DeployMode` take effect locally.