feat: support write wal logs in columnar format (#1179)
## Rationale
After the columnar encoding is supported for wal format, we should adapt
such encoding in the wal write procedure.

## Detailed Changes
Adapt the columnar encoding in the wal write procedure. And now, both
rowwise and columnar encoding in the wal write procedure are supported,
which to use depends on the configuration.

## Test Plan
Existing tests. The compatibility tests have been designed as:
- Write rowwise-encoded wal logs and restart the server to replay it
- Write columnar-encoded wal logs and restart the server to replay it